implemented rate limiter , and properly redirects now
